FSS details improvements and date filtering for chat & game logs (#10)
* Add FSS signals pane and refactor FSS state for better body and signal management. Update journal handling to include signal discovery events.

* Remove FSS signals pane and replace it with a new Body Signals pane. Refactor component macros and scroll handling. Introduce `scroll_list` utility and update version to version 0.7.0. Simplify body data structures and enhance scan updates.

* Refactor GUI components and layout: unify scrollable list handling with `scroll_list`, introduce `bordered_list_item` macro, and simplify alignment/padding logic across panes.

* Refactor state modules: simplify struct and module naming, consolidate imports, and update GUI pane references for improved clarity and maintainability.

* Update FSS body storage and details: consolidate field naming, reduce IDs to `u8` for efficiency, and update GUI and scanning logic accordingly.

* Document utility functions and macros for GUI development: add comprehensive comments and examples for `sub_header`, `details`, `empty_placeholder`, `scroll_list`, and associated macros to improve clarity and developer experience.

* Update journal directory handling and GUI layout: standardize function naming, optimize height/size settings, and simplify directory access and error handling.

* Refactor FSS state to support multiple system scans: replace `fss` with `system_scans`, update body and signal handling logic, and adapt GUI components accordingly.

* Update body name color in Body Signals pane

* Refactor FSS state to support multiple system scans: replace `fss` with `system_scans`, update body and signal handling logic, and adapt GUI components accordingly.

Also- added a setting to allow to filter messages and logs by a certain date.

+++ src/edsm/bodies.rs
@@ -3,21 +3,21 @@ use std::collections::HashMap;
 
 #[derive(Clone, Debug, Deserialize)]
 pub struct Bodies {
-    pub id: Option<i64>,
-    pub id64: Option<i64>,
+    pub id: Option<u32>,
+    pub id64: Option<u64>,
     pub name: Option<Box<str>>,
     pub url: Option<Box<str>>,
     #[serde(rename = "bodyCount")]
-    pub body_count: Option<Option<i32>>,
+    pub body_count: Option<Option<u8>>,
     pub bodies: Option<Vec<Body>>,
 }
 
 #[derive(Clone, Debug, Deserialize)]
 pub struct Body {
-    pub id: i64,
-    pub id64: i64,
+    pub id: u64,
+    pub id64: u64,
     #[serde(rename = "bodyId")]
-    pub body_id: i32,
+    pub body_id: u32,
     pub name: Box<str>,
     pub discovery: Option<Discovery>,
     #[serde(rename = "type")]
